@@ -5,8 +5,8 @@ from contextlib import suppress
 
 http_schemas = ('http', 'https')
 websocket_schemas = ('ws', 'wss')
-http_to_ws = {k: v for k, v in zip(http_schemas, websocket_schemas)}  # noqa: C416
-ws_to_http = {k: v for k, v in zip(websocket_schemas, http_schemas)}  # noqa: C416
+http_to_ws = {k: v for k, v in zip(http_schemas, websocket_schemas, strict=False)}  # noqa: C416
+ws_to_http = {k: v for k, v in zip(websocket_schemas, http_schemas, strict=False)}  # noqa: C416
 
 
 def replace_scheme(url: str, ws: bool) -> str:
